# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Brinecrypt API
  description: |
    Brinecrypt is a secure resource management and encryption service designed to provide a centralized and secure way to store and manage sensitive configuration data, secrets, and other resources.

    ### Core Concepts

    *   **Resources:** Named entities stored within a **Namespace**. Resources can be of type `cleartext` (stored as is) or `encrypted` (transparently encrypted/decrypted).
    *   **Namespaces:** Logical groupings for resources. The `_` namespace is reserved for internal system resources like users.
    *   **Versions:** Every time a resource is updated via `PUT`, a new version is created. Brinecrypt maintains a history of all versions.
    *   **Principals:** Entities that can be granted permissions. These include **Users** and **Service Accounts** (SAs).
    *   **Permissions:** Granular access control based on a `ResourcePattern` (e.g., `prod/*`), a `Verb` (`list`, `read`, `write`, `delete`), and an optional expiry time.

    ### Authentication & Authorization

    Brinecrypt supports multiple authentication methods:
    1.  **Session Tokens:** Obtained via `/auth/login`. Used for interactive sessions.
    2.  **Personal Access Tokens (PATs):** Long-lived tokens issued by users for programmatic access.
    3.  **Capability Tokens:** Restricted tokens issued with a specific set of permissions, useful for delegation or narrow-scoped tasks.
...

...

    ### Encryption

...
  version: 1.0.0
servers:
...
paths:
...
  /api/v1/tokens/pat:
    post:
      tags:
        - Tokens
      summary: Issue a Personal Access Token (PAT)
      operationId: issuePAT
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssuePATRequest'
      responses:
        '200':
          description: PAT successfully issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Unauthorized
  /api/v1/tokens/pat/{id}:
    delete:
      tags:
        - Tokens
      summary: Revoke a PAT
      operationId: revokePAT
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint64
      responses:
        '204':
          description: PAT successfully revoked
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: PAT not found
  /api/v1/tokens/capability:
    post:
      tags:
        - Tokens
      summary: Issue a Capability Token with specific permissions
      operationId: issueCapabilityToken
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueCapabilityTokenRequest'
      responses:
        '200':
          description: Capability token successfully issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /api/v1/tokens/capability/{id}:
    delete:
      tags:
        - Tokens
      summary: Revoke a Capability Token
      operationId: revokeCapabilityToken
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint64
      responses:
        '204':
          description: Capability token successfully revoked
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Token not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Verb:
      type: string
      enum: [list, read, write, delete]
      description: Allowed operations on resources.
...
    PermissionEntry:
      type: object
      properties:
        verb:
          $ref: '#/components/schemas/Verb'
        resource_pattern:
          type: string
          example: "default/my-secret"
        expires_at:
          type: string
          format: date-time
      required:
        - verb
        - resource_pattern
...
    IssuePATRequest:
      type: object
      properties:
        expiry:
          type: string
          format: date-time
    IssueCapabilityTokenRequest:
      type: object
      properties:
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionEntry'
      required:
        - permissions
...